feat: 新增MCP服务支持，完善多语言与构建优化
本次提交包含大量功能更新与代码优化：
1. 新增MCP(Model Context Protocol)服务器，支持AI工具通过标准协议控制设备
2. 添加MCP网页端聊天界面，支持自然语言设备控制
3. 新增侧边栏MCP控制面板，显示服务状态与AI连接命令
4. 完善多语言国际化，新增MCP相关翻译文本
5. 修复大量C解码器编译警告，优化代码健壮性
6. 调整MCP默认端口为10110，同步更新前后端配置
7. 添加workflow主题图标，优化UI界面样式
8. 优化Linux平台编译兼容性，修复类型转换与弃用API警告
9. 新增多套MCP协议测试用例与示例脚本
10. 删除无用的主题更新脚本，清理冗余代码

+++ PXView/pv/dialogs/storeprogress.cpp
@@ -315,7 +315,7 @@ void StoreProgress::export_run()
         
         auto &cursor_list = _view->get_cursorList();
         
-        for (int i=0; i<cursor_list.size(); i++){
+        for (int i=0; i<(int)cursor_list.size(); i++){
             //tr
             QString cursor_name = L_S(STR_PAGE_DLG, S_ID(IDS_DLG_CURSOR), "Cursor") + 
                                 QString::number(i+1);
